requirements.txt
py-applescript==1.0.1
pyobjc==4.2.1